rename `Labels` to `labels`
