+++ src/constant/index.ts
@@ -1,5 +1,5 @@
 import * as API from "./api";
-import LABELS from "./Labels";
+import LABELS from "./labels";
 import { LENGTH } from "./length";
 
 export { API, LABELS, LENGTH };
